Make HTTP authn of API path optional and omitted by default
HTTP authentication on /ovirt-engine/api and /api paths was only
required in oVirt < 4.0. oVirt 4.0 was released 3.5 years ago and this
backward-compatibility setting prevents use of mixed authentication
mechanisms for API (username/password for internal users and kerberos
for LDAP ones).

Given that current setting is obsolete for such a long time and that it
limits available features, I decided to make it disabled by default.